requirements.txt
Cython==3.1.3
setuptools==80.9.0
customtkinter==5.2.2
Cython==3.1.3
darkdetect==0.8.0
packaging==25.0
setuptools==80.9.0
customtkinter==5.2.2
Cython==3.1.3
darkdetect==0.8.0
packaging==25.0
pillow==11.3.0
setuptools==80.9.0
customtkinter==5.2.2
Cython==3.1.3
darkdetect==0.8.0
packaging==25.0
pillow==11.3.0
setuptools==80.9.0
